Move most of `output_file` to the codegen module
